caisp1/webscraper.py
import os
import requests
from dotenv import load_dotenv
from bs4 import BeautifulSoup
from openai import OpenAI

load_dotenv()

api_key = os.getenv("OPENAI_API_KEY")

if not api_key:
    raise ValueError("OPENAI_API_KEY not found in .env file")

client = OpenAI(api_key=api_key)


url = "https://obattin.com"
res = requests.get(url)

soup = BeautifulSoup(res.text, "html.parser")
print(soup.get_text())